Replace regular objects with omni objects
